requirements.txt
asgiref==3.5.2
astroid==2.12.10
certifi==2022.9.14
charset-normalizer==2.1.1
dill==0.3.5.1
Django==4.1.1
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.15.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
ujson==5.5.0
urllib3==1.26.12
wrapt==1.14.1
django
djangorestframework
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
celery
django-import-export
redis